Added path number functionality
Now we can continue building the path tree from the 'current_node' (position of Indy), in order to find a new path through the maze "on the fly". Also this allows us to traverse the tree from top to bottom following the active path of movement of Indy (before we could do that only from the very bottom).